User commit

+++ src/main/java/com/example/demo/repository/StudentRepo.java
@@ -1,10 +1,10 @@
 package com.example.demo.repository;
 
-import org. springframework. data. jpa.repository. JpaRepository;
-import org. springframework. stereotype. Repository;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import com.example.demo.entity.StudentEntity;
 
-import com.example.demo.entity.Student;
 @Repository
-public interface StudentRepo extends JpaRepository<StudentEntity, Integer>{
+public interface StudentRepo extends JpaRepository<StudentEntity, Long> {
 
 }
